pyproject.toml
[project]
name = "rag-ollama"
version = "0.1.0"
description = "RAG system with Ollama"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling",
    "langchain",
    "langchain-community",
    "pydantic",
    "chromadb",
    "sentence-transformers",
    "unstructured",
    "langchain-text-splitters",
    "langchain-core",
    "rank_bm25",
    "langchain-chroma",
    "langchain-ollama",
    "markdown",
    "openai",
    "requests",
    "ollama",
    "pdf-ocr-ai",
    "PyYAML",
    "tqdm",
]

[project.scripts]
rag-prepare = "rag_ollama.prepare_documents:main"
rag-chat = "rag_ollama.rag:main"
rag-add = "rag_ollama.add_document:main"
rag-ollama = "rag_ollama.rag:main"

[tool.uv]
dev-dependencies = [
    "pytest",
]

[tool.uv.sources]
pdf-ocr-ai = { git = "https://github.com/laurentvv/pdf-ocr-ai" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
